requirements.txt
matplotlib==3.10.3
mysql_connector_repackaged==0.3.1
numpy==2.2.6
pandas==2.2.3
python-dotenv==1.1.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
